refactor: move rest of the classes out from com.hedera.node.app.spi.state and TestSchema
