+++ hedera-node/hedera-app-spi/build.gradle.kts
@@ -25,7 +25,6 @@ description = "Hedera Application - SPI"
 testModuleInfo {
     requires("com.hedera.node.app.spi")
     requires("com.swirlds.state.api.test.fixtures")
-    requires("com.swirlds.platform.core.test.fixtures")
     requires("org.apache.commons.lang3")
     requires("org.assertj.core")
     requires("org.junit.jupiter.api")
