+++ hedera-node/hedera-app-spi/src/testFixtures/java/com/hedera/node/app/spi/fixtures/TestService.java
@@ -19,10 +19,10 @@ package com.hedera.node.app.spi.fixtures;
 import static java.util.Objects.requireNonNull;
 
 import com.hedera.node.app.hapi.utils.ethereum.EthTxData;
-import com.swirlds.platform.test.fixtures.state.TestSchema;
 import com.swirlds.state.lifecycle.Schema;
 import com.swirlds.state.lifecycle.SchemaRegistry;
 import com.swirlds.state.lifecycle.Service;
+import com.swirlds.state.test.fixtures.merkle.TestSchema;
 import edu.umd.cs.findbugs.annotations.NonNull;
 import java.util.ArrayList;
 import java.util.List;
