+++ hedera-node/hedera-app/src/test/java/com/hedera/node/app/services/ServiceRegistryImplTest.java
@@ -27,8 +27,8 @@ import com.hedera.hapi.node.base.Timestamp;
 import com.hedera.node.app.spi.fixtures.TestService;
 import com.swirlds.common.constructable.ConstructableRegistry;
 import com.swirlds.common.constructable.ConstructableRegistryException;
-import com.swirlds.platform.test.fixtures.state.TestSchema;
 import com.swirlds.state.lifecycle.StateDefinition;
+import com.swirlds.state.test.fixtures.merkle.TestSchema;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
