+++ hedera-node/hedera-token-service-impl/build.gradle.kts
@@ -31,7 +31,7 @@ testModuleInfo {
     requires("com.hedera.node.app.service.token.test.fixtures")
     requires("com.swirlds.state.api.test.fixtures")
     requires("com.swirlds.config.extensions.test.fixtures")
-    requires("com.swirlds.platform.core.test.fixtures")
+    requires("com.swirlds.state.impl.test.fixtures")
     requires("net.i2p.crypto.eddsa")
     requires("org.assertj.core")
     requires("org.junit.jupiter.api")
