+++ platform-sdk/swirlds-state-impl/src/testFixtures/java/module-info.java
@@ -4,6 +4,7 @@ open module com.swirlds.state.impl.test.fixtures {
     requires transitive com.swirlds.config.api;
     requires transitive com.swirlds.merkle;
     requires transitive com.swirlds.state.api.test.fixtures;
+    requires transitive com.swirlds.state.api;
     requires transitive com.swirlds.state.impl;
     requires transitive com.swirlds.virtualmap;
     requires transitive com.hedera.pbj.runtime;
